Moved excel writer to repn/plugins
