+++ pyomo/repn/plugins/__init__.py
@@ -14,4 +14,5 @@ def load():
     import pyomo.repn.plugins.baron_writer
     import pyomo.repn.plugins.mps
     import pyomo.repn.plugins.gams_writer
+    import pyomo.repn.plugins.excel_writer
